# #1727 - Use faster disambiguation in isdisjoint(::CPA, ::HalfSpace)