Closes #1727.

Platooning benchmark:

`master`:
          ("PLAD01_BND42", "discrete") => TrialEstimate(282.038 ms)
          ("PLAN01_UNB50", "discrete") => TrialEstimate(7.917 s)
          ("PLAD01_BND42", "dense") => TrialEstimate(860.151 ms)
          ("PLAN01_UNB50", "dense") => TrialEstimate(11.131 s)

this branch:
          ("PLAD01_BND42", "discrete") => TrialEstimate(204.435 ms)
          ("PLAN01_UNB50", "discrete") => TrialEstimate(7.312 s)
          ("PLAD01_BND42", "dense") => TrialEstimate(699.705 ms)
          ("PLAN01_UNB50", "dense") => TrialEstimate(10.025 s)

(Note that the performance is not as good as before the commit mentioned in #1727 because there is another regression.)